requirements.txt
astapi
uvicorn
scikit-learn
joblib
pandas
numpy